Refactor file stucture for fetching
