--- src/helpers/__mocks__/createNewUser.js
@@ -1,4 +1,0 @@
-const mockReturnedUser = {
-  id: 5
-}
-export const createNewUser = jest.fn().mockImplementation(() => Promise.resolve(mockReturnedUser));
