--- src/helpers/__mocks__/getUser.js
@@ -1,8 +1,0 @@
-const mockUser = {
-    data: {
-        id: 2,
-        name: 'John'
-    }
-}
-
-export const getUser = jest.fn().mockImplementation(() => Promise.resolve(mockUser))
